@@ -8,36 +8,39 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^3.9.0",
-    "@material-ui/icons": "^3.0.2",
+    "@craco/craco": "^5.2.1",
+    "@material-ui/core": "^4.0.2",
+    "@material-ui/icons": "^4.0.1",
+    "@material-ui/styles": "^4.0.2",
     "classnames": "^2.2.5",
     "ease-component": "^1.0.0",
     "parse-link-header": "^1.0.1",
     "pusher-js": "^4.2.2",
     "qs": "^6.5.1",
-    "ramda": "^0.25.0",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-redux": "^5.0.6",
+    "ramda": "^0.26.1",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.0.3",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.3",
     "recompose": "^0.30.0",
-    "redux": "^3.7.2",
-    "redux-observable": "^0.18.0",
-    "rxjs": "^5.4.3",
+    "redux": "^4.0.1",
+    "redux-observable": "^1.1.0",
+    "rxjs": "^6.5.2",
     "scroll": "^2.0.1",
-    "scroll-doc": "^0.2.0"
+    "scroll-doc": "^0.2.0",
+    "webpack-bundle-analyzer": "^3.3.2"
   },
   "devDependencies": {
     "jest-fetch-mock": "^1.2.1",
     "redux-logger": "^3.0.6"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "start": "craco start",
+    "build": "craco build",
+    "analyze": "ANALYZE=yes craco build",
+    "test": "craco test"
   },
   "eslintConfig": {
     "extends": "react-app"
