@@ -1,12 +1,14 @@
-import defaultTo from 'ramda/src/defaultTo';
-import converge from 'ramda/src/converge';
-import pair from 'ramda/src/pair';
-import partial from 'ramda/src/partial';
-import ifElse from 'ramda/src/ifElse';
-import compose from 'ramda/src/compose';
-import prop from 'ramda/src/prop';
-import nth from 'ramda/src/nth';
-import isNil from 'ramda/src/isNil';
+import {
+    defaultTo,
+    converge,
+    pair,
+    partial,
+    ifElse,
+    compose,
+    prop,
+    nth,
+    isNil,
+} from 'ramda';
 import parseLinkHeader from 'parse-link-header';
 
 import allP from './allP';
