@@ -1,7 +1,9 @@
-import map from 'ramda/src/map';
-import flatten from 'ramda/src/flatten';
-import compose from 'ramda/src/compose';
-import prop from 'ramda/src/prop';
+import {
+    map,
+    flatten,
+    compose,
+    prop,
+ } from 'ramda';
 import allP from './allP';
 import fetchUserInstallations from './fetchUserInstallations';
 import fetchInstallationRepositories from './fetchInstallationRepositories';
