@@ -1,6 +1,8 @@
-import find from 'ramda/src/find';
-import propEq from 'ramda/src/propEq';
-import defaultTo from 'ramda/src/defaultTo';
+import {
+    find,
+    propEq,
+    defaultTo,
+ } from 'ramda';
 
 export default (pullRequestNumber, queue) =>
     defaultTo({}, find(propEq('pullRequestNumber', pullRequestNumber), queue));
