@@ -1,4 +1,3 @@
-import curry from 'ramda/src/curry';
-import map from 'ramda/src/map';
+import { curry, map } from 'ramda';
 
 export default curry((keys, source) => map(key => source[key], keys));
