@@ -1,3 +1,3 @@
-import curry from 'ramda/src/curry';
+import { curry } from 'ramda';
 
 export default curry((length, string) => string.substr(length));
