@@ -1,9 +1,10 @@
-import prop from 'ramda/src/prop';
-import propEq from 'ramda/src/propEq';
-import compose from 'ramda/src/compose';
-import startsWith from 'ramda/src/startsWith';
-import path from 'ramda/src/path';
-
+import {
+    prop,
+    propEq,
+    compose,
+    startsWith,
+    path,
+} from 'ramda';
 export const requestIdEq = requestId => compose(
     propEq('requestId', requestId),
     prop('payload'),
